현우 주문, 결제 브라우저 종료시 삭제 commit

+++ Jungkosta/src/main/java/jungkosta/auction/service/AucPurServiceImpl.java
@@ -14,16 +14,17 @@ import jungkosta.auction.domain.BiddingVO;
 import jungkosta.auction.persistence.AucPurDAO;
 import jungkosta.auction.persistence.AuctionDAO;
 import jungkosta.auction.persistence.BiddingDAO;
+import jungkosta.commons.util.AuctionCheck;
 
 @Service
 public class AucPurServiceImpl implements AucPurService {
 
 	@Inject
 	private AucPurDAO dao;
-	
+
 	@Inject
 	private BiddingDAO biddingDao;
-	
+
 	@Inject
 	private AuctionDAO auctionDao;
 
@@ -33,11 +34,13 @@ public class AucPurServiceImpl implements AucPurService {
 		vo.setBid_id(dao.selectBid_id() + 1);
 		vo.setBid_status("입금 대기");
 		dao.insertAuctionPurchase(vo);
-		
+
 		BiddingVO bidding = biddingDao.readBidding(vo.getBidding_id());
-		
+
 		dao.updateSale_status(sale_id);
 		dao.endAuction(bidding.getAuction_id());
+		
+		System.out.println("주문 완료");
 	}
 
 	@Override
@@ -55,16 +58,18 @@ public class AucPurServiceImpl implements AucPurService {
 	@Transactional
 	@Override
 	public void cancelPurchase(int sale_id, int bidding_id) throws Exception {
-		
 		AuctionVO auction = auctionDao.read(sale_id);
-		
+
 		Map<String, Integer> map = new HashMap<>();
-		
+
 		map.put("sale_id", sale_id);
 		map.put("auction_id", auction.getAuction_id());
+		if (AuctionCheck.compareTime(auction.getAuction_end_date())) {
+			dao.deleteBidding(bidding_id);
+			dao.updateSale_cost(map);
+		}
 		
-		dao.deleteBidding(bidding_id);
-		dao.updateSale_cost(map);
+		System.out.println("주문 취소");
 	}
 
 }
